backend/requirements.txt
Django>=5.0,<6.0
dj-database-url>=2.2,<3.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
gunicorn>=22.0,<23.0
openpyxl>=3.1,<4.0
Pillow>=10.0,<11.0
psycopg[binary]>=3.2,<4.0
whitenoise>=6.7,<7.0
